Base start up
CLEO can start a new file now. When '-wineoln' is set, the file will be created with a DOS based EOLN which is CRLF, otherwise it will be done the Unix way LF.
In files loaded CLEO will try to autodetect and set it to the DOS based EOLN if CRLF is detected. Of course, if the file has been set up inconsistently CLEO cannot function normally, but inconsistent EOLNs are a sin anyway.